requirements.txt


# this is the "requirements.txt" file...


requests

plotly

python-dotenv

sendgrid==6.5.0

pandas

pytest

flask

gunicorn

IPython

pgeocode

